refactor: extract gateway execute module
to make it easier to write isolated unit tests I'm extracting the module into its own file
